--- src/algorithm/index.ts
@@ -1,3 +1,0 @@
-export * from './apply';
-export * from './blueprint';
-export * from './parser';
